A Better File Structure Done
