+++ controllers/userController.js
@@ -1,0 +1,34 @@
+
+const getAllUsers = (req, res) => {
+    res.status(500).json({
+        status: "error",
+        message: "route for this endpoint is not defined yet"
+    })
+}
+const createAUser = (req, res) => {
+    res.status(500).json({
+        status: "error",
+        message: "route for this endpoint is not defined yet"
+    })
+}
+const getAUser = (req, res) => {
+    res.status(500).json({
+        status: "error",
+        message: "route for this endpoint is not defined yet"
+    })
+}
+const updateAUser = (req, res) => {
+    res.status(500).json({
+        status: "error",
+        message: "route for this endpoint is not defined yet"
+    })
+}
+const deleteAUser = (req, res) => {
+    res.status(500).json({
+        status: "error",
+        message: "route for this endpoint is not defined yet"
+    })
+}
+
+
+export { getAUser, getAllUsers, createAUser, deleteAUser, updateAUser }
