+++ routers/route.js
@@ -1,0 +1,4 @@
+import tourRouter from './tourRoutes.js'
+import userRouter from './userRoutes.js'
+
+export { tourRouter, userRouter }
