+++ server.js
@@ -1,0 +1,7 @@
+import app from "./app.js"
+const PORT = 5000;
+
+
+app.listen(PORT, () => {
+    console.log(`server is started http://localhost:${PORT}`);
+})
